+++ packages/stylelint/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## stylelint@0.0.3 (2026-02-23)
+
+- Add `declaration-property-value-no-unknown` rule for miniapp
+
 ## stylelint@0.0.2 (2026-02-20)
 
 - Update `exports` field for CommonJS compatibility